+++ garage/baselines/gaussian_conv_baseline.py
@@ -4,7 +4,7 @@ from garage.baselines import Baseline
 from garage.core import Parameterized
 from garage.core import Serializable
 from garage.misc.overrides import overrides
-from garage.regressors import GaussianConvRegressor
+from garage.theano.regressors import GaussianConvRegressor
 
 
 class GaussianConvBaseline(Baseline, Parameterized):
