+++ garage/theano/regressors/__init__.py
@@ -1,0 +1,6 @@
+from garage.theano.regressors.categorical_mlp_regressor import \
+        CategoricalMLPRegressor
+from garage.theano.regressors.gaussian_conv_regressor import \
+        GaussianConvRegressor
+from garage.theano.regressors.gaussian_mlp_regressor import \
+        GaussianMLPRegressor
