+++ garage/theano/regressors/categorical_mlp_regressor.py
