+++ garage/theano/regressors/gaussian_conv_regressor.py
