+++ garage/theano/regressors/gaussian_mlp_regressor.py
